Functionality Added to Resolve Page
Double click on a row in the service request view page to get to the resolve page.
Uses a singleton method to get data from rows to the resolve page controller.
All fields that can be filled in on resolve page are. Database query is used for this.
Status is updated when confirm is pressed.